+++ BASE/Test/ServerTelemetryChannel.Test/TelemetryChannel.Tests/Implementation/ErrorHandlingTransmissionPolicyTest.cs
@@ -214,7 +214,7 @@ namespace Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.Implement
                     Assert.AreEqual(69, traces[0].EventId); // failed to send
                     Assert.AreEqual("Data loss", traces[0].Payload[1]);
                     // Sets flush task to failure on not whitelisted status code
-                    Assert.IsFalse(failedTransmission.HasFlushTask);
+                    Assert.IsFalse(failedTransmission.IsFlushAsyncInProgress);
                 }
             }
 
@@ -247,7 +247,7 @@ namespace Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.Implement
                     Assert.AreEqual(69, traces[0].EventId); // failed to send
                     Assert.AreEqual("Unknown Exception Message", traces[0].Payload[1]);
                     // Sets flush task to failure on not whitelisted status code
-                    Assert.IsFalse(failedTransmission.HasFlushTask);
+                    Assert.IsFalse(failedTransmission.IsFlushAsyncInProgress);
                 }
             }
 
@@ -284,7 +284,7 @@ namespace Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.Implement
                     Assert.AreEqual(71, traces[0].EventId); // failed to send
                     Assert.AreEqual("8989", traces[0].Payload[1]);
                     // Sets flush task to failure on not whitelisted status code
-                    Assert.IsFalse(failedTransmission.HasFlushTask);
+                    Assert.IsFalse(failedTransmission.IsFlushAsyncInProgress);
                 }
             }